+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+# 1.3.0
+* \+ : deviantArt - Ignore "tier" notification
+* \+ : Native messaging - Open url in another browser
+* Fix : Change deviantArt favicon url 
+
 # 1.2.0
 * \+ : Native messaging - Change port id and code cleaning
-* 
+
 # 1.1.0
 * i : Move from Mustache templates to Twig using [twig.js](https://github.com/twigjs/twig.js)
 * i : Add DNSlytics link with Page IP / Status display